requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12
httpx>=0.27.0
pymupdf>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10
